+++ nv/plugin_unimpaired.py
@@ -25,6 +25,8 @@ from NeoVintageous.nv.plugin import NORMAL
 from NeoVintageous.nv.plugin import register
 from NeoVintageous.nv.plugin import ViOperatorDef
 from NeoVintageous.nv.plugin import VISUAL
+from NeoVintageous.nv.window import window_buffer_control
+from NeoVintageous.nv.window import window_tab_control
 
 
 __all__ = [
@@ -84,6 +86,58 @@ class _UnimpairedBlankDown(ViOperatorDef):
         }
 
 
+@register(seq='[b', modes=(NORMAL,))
+class _UnimpairedBprevious(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'bprevious'
+            }
+        }
+
+
+@register(seq=']b', modes=(NORMAL,))
+class _UnimpairedBnext(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'bnext'
+            }
+        }
+
+
+@register(seq='[B', modes=(NORMAL,))
+class _UnimpairedBfirst(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'bfirst'
+            }
+        }
+
+
+@register(seq=']B', modes=(NORMAL,))
+class _UnimpairedBlast(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'blast'
+            }
+        }
+
+
 @register(seq='[e', modes=(NORMAL,))
 class _UnimpairedMoveUp(ViOperatorDef):
     def translate(self, state):
@@ -106,6 +160,58 @@ class _UnimpairedMoveDown(ViOperatorDef):
                 'mode': state.mode,
                 'count': state.count,
                 'action': 'move_down'
+            }
+        }
+
+
+@register(seq='[t', modes=(NORMAL,))
+class _UnimpairedTabprevious(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'tabprevious'
+            }
+        }
+
+
+@register(seq=']t', modes=(NORMAL,))
+class _UnimpairedTabnext(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'tabnext'
+            }
+        }
+
+
+@register(seq='[T', modes=(NORMAL,))
+class _UnimpairedTabfirst(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'tabfirst'
+            }
+        }
+
+
+@register(seq=']T', modes=(NORMAL,))
+class _UnimpairedTablast(ViOperatorDef):
+    def translate(self, state):
+        return {
+            'action': '_nv_unimpaired',
+            'action_args': {
+                'mode': state.mode,
+                'count': state.count,
+                'action': 'tablast'
             }
         }
 
@@ -400,6 +506,10 @@ class _nv_unimpaired_command(TextCommand):
         elif action == 'blank_up':
             # Add [count] blank lines above the cursor
             _blank_up(self.view, edit, count)
+        elif action in ('bnext', 'bprevious', 'bfirst', 'blast'):
+            window_buffer_control(self.view.window(), action[1:], count=count)
+        elif action in ('tabnext', 'tabprevious', 'tabfirst', 'tablast'):
+            window_tab_control(self.view.window(), action[3:], count=count)
         elif action == 'context_next':
             # Go to the next [count]  SCM conflict marker or diff/patch hunk
             _context_next(self.view, count)